requirements.txt
# Python packages
time
os

# Selenium package
selenium

# Pillow package
Pillow

# Keras package
keras
numpy
